+++ src/Xml/Dsig/ArrayReferenceResolver.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Allkiri\Xades\Dsig;
+namespace Allkiri\Xml\Dsig;
 
 final class ArrayReferenceResolver implements ReferenceResolver
 {